refactor(view): 수익률 계산에 대한 소수점 정밀화
소수점 출력이 현재 정확하게 나오지 않고 소수점 2자리까지 출력이 되지 않는 문제가 발생하였습니다.

BigDecimal을 사용하여 계산된 결과를 View에 전달하여 출력되도록 수정하였습니다.

+++ src/main/java/lotto/view/OutputView.java
@@ -2,6 +2,7 @@ package lotto.view;
 
 import lotto.domain.result.RankResult;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class OutputView {
@@ -28,7 +29,7 @@ public class OutputView {
     }
 
     // 수익률 출력
-    public void printYield(double yield) {
+    public void printYield(BigDecimal yield) {
         System.out.println("총 수익률은 " + yield + "%입니다.");
     }
